StreamBuilderをListのwidgtsに置換
